Separate summary items into files
