+++ Neo4j.Driver/Neo4j.Driver/Internal/Result/Summary/IUpdateableInfo.cs
@@ -1,0 +1,8 @@
+﻿using Neo4j.Driver.Internal.Protocol;
+
+namespace Neo4j.Driver.Internal.Result;
+
+internal interface IUpdateableInfo
+{
+    void Update(BoltProtocolVersion boltVersion, string agent);
+}
